Adds Support for virtual method overrides within lua.
Very, Very *VERY* rough work in progress on an approach to allowing
implementation of the virtual method overrides of the 'standard'
Cocos2dx objects.

So many risky things here, but threads springs to mind immediately
there is still no LUA_LOCK concept to lock on to avoid concurrency,
but as threading is currently disabled in cocos2d-x that shouldn't be an issue.

Cannot currently (or foreseeably?) override the init method (frustrating as
that is what I mostly did this for! ) as the CCLayer:node() idiom returns
and already constructed object before we've had chance to configure the
metatable for doing the vtable looky-likey :( ... workaroundable certainly
but is it worth it ?